requirements.txt
tensorflow
tensorflow_datasets
matplotlib
numpy
rich
pyyaml